Experiment 4 run2: DeBERTa scoring on condA and condC
- scores/free_gen_qwen7b_metaft_run2_condA_score.json  -- condA scored
- scores/free_gen_qwen7b_metaft_run2_condC_score.json  -- condC scored
- free_gen_triage.py / filter_balance_clean.py         -- missing support modules added
- train.tsv / val.tsv                                  -- classifier training data (407/71 pairs)
- train_deberta.py / run_deberta.py                    -- fp32 fix for DeBERTa-v3 NaN issue

DeBERTa trained on run1 condA+C clean examples; val AUROC 0.992
CondA detection rate: 0.998  CondC detection rate: 0.994

+++ chronologic/bertclassify/run_deberta.py
@@ -284,7 +284,7 @@ def main(argv=None):
     # Load model + tokenizer
     print(f"\nLoading model from {model_dir} ...")
     tokenizer = AutoTokenizer.from_pretrained(str(model_dir))
-    model = AutoModelForSequenceClassification.from_pretrained(str(model_dir))
+    model = AutoModelForSequenceClassification.from_pretrained(str(model_dir), torch_dtype=torch.float32)
     model = model.to(device)
     print(f"Model loaded ({sum(p.numel() for p in model.parameters()):,} parameters)")
